+++ src/LVGLSharp.Runtime.Remote/IRemoteFrameSource.cs
@@ -1,0 +1,6 @@
+namespace LVGLSharp.Runtime.Remote;
+
+public interface IRemoteFrameSource
+{
+    RemoteFrame CaptureFrame();
+}